+++ global_helpers/default_test.py
@@ -6,11 +6,11 @@ import sys
 import unittest
 
 sys.path.append(os.path.dirname(__file__))
-import panther_aws_helpers as p_d  # pylint: disable=C0413
+import panther_aws_helpers as p_aws_h  # pylint: disable=C0413
 
 
 class TestAWSKeyAccountId(unittest.TestCase):
     def test_aws_key_account_id(self):
         aws_key_id = "ASIAY34FZKBOKMUTVV7A"
-        account_id = p_d.aws_key_account_id(aws_key_id)
+        account_id = p_aws_h.aws_key_account_id(aws_key_id)
         self.assertEqual(account_id, "609629065308")
