+++ global_helpers/panther_gcp_helpers.py
@@ -59,7 +59,6 @@ def get_binding_deltas(event):
     if not service_data:
         return []
 
-    # Reference: bit.ly/2WsJdZS
     binding_deltas = service_data.get("policyDelta", {}).get("bindingDeltas")
     if not binding_deltas:
         return []
