--- global_helpers/panther_oss_helpers.yml
@@ -1,4 +1,0 @@
-AnalysisType: global
-GlobalID: "panther_oss_helpers"
-Filename: panther_oss_helpers.py
-Description: Used to define global helpers and variables.
