+++ packs/aws.yml
@@ -198,4 +198,3 @@ PackDefinition:
     - panther_iocs
     - panther_ipinfo_helpers
     - panther_lookuptable_helpers
-    - panther_oss_helpers
