+++ packs/aws_cis.yml
@@ -37,7 +37,7 @@ PackDefinition:
     - AWS.WAF.HasXSSPredicate
     # Globals used in these detections
     - panther_base_helpers
-    - panther_oss_helpers
+    - panther_aws_helpers
     - panther_config
     - panther_config_defaults
     - panther_config_overrides
