+++ packs/standard_ruleset.yml
@@ -35,4 +35,3 @@ PackDefinition:
     - panther_event_type_helpers
     - panther_ipinfo_helpers
     - panther_lookuptable_helpers
-    - panther_oss_helpers
