+++ policies/aws_iam_policies/aws_access_key_unused.py
@@ -1,6 +1,6 @@
 import datetime
 
-from panther_oss_helpers import resolve_timestamp_string
+from panther_base_helpers import resolve_timestamp_string
 
 TIMEOUT_DAYS = datetime.timedelta(days=90)
 DEFAULT_TIME = "0001-01-01T00:00:00Z"
