+++ rules/standard_rules/brute_force_by_ip.py
@@ -2,8 +2,8 @@ from json import loads
 
 import panther_event_type_helpers as event_type
 from panther_aws_helpers import lookup_aws_account_name
+from panther_base_helpers import add_parse_delay
 from panther_ipinfo_helpers import PantherIPInfoException, geoinfo_from_ip
-from panther_oss_helpers import add_parse_delay
 
 
 def rule(event):
